refactor: renderSelector 모듈이 파일 내 모든 selector에 적용하도록 변경
